requirements.txt
# Core dependencies
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pydantic>=2.0.0,<3.0.0

# LlamaIndex and related packages
llama-index>=0.11.0
llama-index-llms-google-genai>=0.1.0
llama-index-embeddings-huggingface>=0.5.0
langchain-core>=0.1.0,<0.3.0
langchain-community>=0.0.30,<0.1.0

# REST API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

# Machine learning and embeddings
torch>=2.0.0
transformers>=4.41.0
sentence-transformers>=2.6.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Additional utilities
huggingface-hub>=0.19.0
filelock>=3.12.0
tokenizers>=0.21.0
safetensors>=0.4.0